Review fixes: renames
